Renamed helper DataStructure => CaseDataConverter
