+++ source/helpers/index.js
@@ -1,5 +1,5 @@
 import * as ApiRequest from './ApiRequest';
-import * as DataStructure from './DataStructure';
+import * as CaseDataConverter from './CaseDataConverter';
 import * as EventHandler from './EventHandler';
 import * as MarkdownConstructor from './MarkdownConstructor';
 import * as MessageHelper from './MessageHelper';
@@ -9,7 +9,7 @@ import * as ValidationHelper from './ValidationHelper';
 
 export {
   ApiRequest,
-  DataStructure,
+  CaseDataConverter,
   EventHandler,
   MarkdownConstructor,
   MessageHelper,
